requirements.txt
Django==2.2.24
django-bulk-update==2.2.0
djangorestframework==3.12.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
